recovery: SEARCH should work as SELECT_ITEM
Before 2cc22971e40a0ffc7849cdf09dc96916c22136e0, SEARCH worked as
SELECT_ITEM. After that, it became "back".

It is desirable for SEARCH to work as select because it provides
a full control for devices with the 4-capacitive-buttons row.

MENU = DOWN
HOME = UP
SEARCH = SELECT (was BACK)
BACK = BACK

Change-Id: Idc70366d76a18edecb25adce2e35e31e814443c5